Restructured the agent definitions.
