+++ .vscode/settings.json
@@ -9,6 +9,7 @@
     "dunder",
     "Fira",
     "Fullscreen",
+    "funcs",
     "geteuid",
     "getpid",
     "journaling",
