+++ agents_playground/agents/default/default_agent_movement_attributes.py
@@ -1,0 +1,6 @@
+from agents_playground.agents.spec.agent_movement_attributes import AgentMovementAttributes
+
+class DefaultAgentMovementAttributes(AgentMovementAttributes):
+  """By default there aren't any movement specific attributes"""
+  def __init__(self) -> None:
+    ...
