+++ agents_playground/agents/default/map_agent_action_selector_exception.py
@@ -1,0 +1,3 @@
+class MapAgentActionSelectorException(Exception):
+  def __init__(self, *args: object) -> None:
+    super().__init__(*args)
