+++ agents_playground/agents/default/named_agent_state.py
@@ -1,0 +1,5 @@
+from agents_playground.agents.spec.agent_state_spec import AgentActionStateLike
+
+class NamedAgentState(AgentActionStateLike):
+  def __init__(self, name: str) -> None:
+    self.name = name
