+++ agents_playground/agents/spec/agent_action_state_spec.py
@@ -1,0 +1,4 @@
+from typing import Protocol
+
+class AgentActionStateLike(Protocol):
+  name: str
