+++ agents_playground/agents/spec/agent_life_cycle_phase.py
@@ -1,0 +1,6 @@
+from enum import Enum, auto
+
+
+class AgentLifeCyclePhase(Enum):
+  PRE_STATE_CHANGE  = auto()
+  POST_STATE_CHANGE = auto()
