+++ agents_playground/agents/spec/agent_movement_attributes.py
@@ -1,0 +1,8 @@
+from typing import Protocol
+
+
+class AgentMovementAttributes(Protocol):
+  """
+  Responsible for storing any agent specific attributes related to movement.
+  """
+  ...
