+++ demo/a_star_navigation/a_star_navigation/agent_states.py
@@ -1,21 +1,21 @@
-from agents_playground.agents.agent_spec import AgentMovementAttributes
-from agents_playground.agents.default_agent import NamedAgentState
+from agents_playground.agents.spec.agent_spec import AgentMovementAttributes
+from agents_playground.agents.default.default_agent import NamedAgentState
 from agents_playground.counter.counter import Counter
 from agents_playground.paths.interpolated_path import InterpolatedPath
 
-IDLE_STATE = NamedAgentState('IDLE')
-RESTING_STATE = NamedAgentState('RESTING')
-PLANNING_STATE = NamedAgentState('PLANNING')
-ROUTING_STATE = NamedAgentState('ROUTING')
-TRAVELING_STATE = NamedAgentState('TRAVELING')
+# IDLE_STATE = NamedAgentState('IDLE')
+# RESTING_STATE = NamedAgentState('RESTING')
+# PLANNING_STATE = NamedAgentState('PLANNING')
+# ROUTING_STATE = NamedAgentState('ROUTING')
+# TRAVELING_STATE = NamedAgentState('TRAVELING')
 
-AgentStateMap = {
-  IDLE_STATE: IDLE_STATE,
-  RESTING_STATE: PLANNING_STATE,
-  PLANNING_STATE: ROUTING_STATE,
-  ROUTING_STATE: TRAVELING_STATE,
-  TRAVELING_STATE: RESTING_STATE
-}
+# AgentStateMap = {
+#   IDLE_STATE: IDLE_STATE,
+#   RESTING_STATE: PLANNING_STATE,
+#   PLANNING_STATE: ROUTING_STATE,
+#   ROUTING_STATE: TRAVELING_STATE,
+#   TRAVELING_STATE: RESTING_STATE
+# }
 
 class PathConstrainedAgentMovement(AgentMovementAttributes):
   resting_counter: Counter
