+++ tests/agents/agent_position_test.py
@@ -1,8 +1,7 @@
 
-from agents_playground.agents.default_agent import DefaultAgentPosition
+from agents_playground.agents.default.default_agent_position import DefaultAgentPosition
 from agents_playground.agents.direction import Vector2d
 from agents_playground.core.types import Coordinate
-
 
 class TestAgentPosition:
   def test_agent_position(self) -> None:
